requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
email-validator
Werkzeug
requests
python-telegram-bot
Flask-Migrate
python-telegram-bot
